feat: finish number and fiat number display componnt
