+++ packages/widgets-internal/components/NumberDisplay/index.ts
@@ -1,0 +1,2 @@
+export * from "./FiatNumberDisplay";
+export * from "./NumberDisplay";
